update: views.py models.py
models.py: added a 'special_state' field to the game model that is going to indicate states like 'battle' or 'demand' and let us know how long has it been going on for (in the case of a battle). views.py: updated the state view in order to return proper top_card when no moves have been made yet. Started working on the move view